Reface: split to options file
