+++ src/lib.rs
@@ -3,5 +3,6 @@ pub mod detections;
 pub mod filter;
 pub mod notify;
 pub mod omikuji;
+pub mod options;
 pub mod timeline;
 pub mod yaml;
